+++ mdsshr/MdsGet1DxS.c
@@ -49,13 +49,13 @@ STATIC_CONSTANT char *cvsrev = "@(#)$RCSfile$ $Revision$ $Date$";
 
 #define align(bytes,size) ((((bytes) + (size) - 1)/(size)) * (size))
 
-int       MdsGet1DxS(unsigned short *length_ptr, unsigned char *dtype_ptr, struct descriptor_xd *out_dsc_ptr)
+int       MdsGet1DxS(unsigned long *length_ptr, unsigned char *dtype_ptr, struct descriptor_xd *out_dsc_ptr)
 {
 
   int       status;
   int dsc_size = sizeof(struct descriptor);
   int align_size = (*dtype_ptr == DTYPE_T) ? 1 : *length_ptr;
-  unsigned int length;
+  unsigned long length;
   STATIC_CONSTANT unsigned char dsc_dtype = DTYPE_DSC;
   dsc_size = align(dsc_size,align_size);
   length = dsc_size + *length_ptr;
