+++ tdishr/TdiArray.c
@@ -46,7 +46,7 @@ array_coeff arr = {1,DTYPE_B,CLASS_A,(char *)0,0,0,{0,1,1,1,0},MAXDIM,0};
 array_int cvt = {sizeof(int),DTYPE_L,CLASS_A,(int *)0,0,0,{0,1,1,0,0},1,0};
 struct TdiFunctionStruct	*fun_ptr = (struct TdiFunctionStruct *)&TdiRefFunction[opcode];
 struct descriptor_xd	tmp = EMPTY_XD;
-unsigned short length;
+unsigned long length;
 unsigned char dtype;
 int				j, ndim=0;
