+++ tdishr/TdiSubscript.c
@@ -233,7 +233,7 @@ struct TdiCatStruct		cats[2];
 	if (psig && psig->ndesc > highdim + 2) psig->ndesc = (unsigned char)(highdim + 2);
 	if (psig) for (j = psig->ndesc; --j >= 0;) psig->dimensions[j] = xx[j].pointer;
 	if (highest <= 0) {
-                unsigned short llen = (unsigned short)0;
+                unsigned long llen = 0;
                 unsigned char dtype = (unsigned char)DTYPE_MISSING;
 		if (arr.arsize) status = MdsGet1DxS(&pdat->length, &pdat->dtype, out_ptr);
 		else status = MdsGet1DxS(&llen, &dtype, out_ptr);
