+++ tdishr/tdirefcat.h
@@ -15,15 +15,15 @@
 #define TdiCAT_D		(0x8000 | TdiCAT_FLOAT | 7)
 #define TdiCAT_FC		(TdiCAT_COMPLEX | TdiCAT_F)
 struct TdiCatStruct {
-	unsigned short	in_cat;
+	unsigned long	in_cat;
 	unsigned char	in_dtype;
-	unsigned short	out_cat;
+	unsigned long	out_cat;
 	unsigned char	out_dtype;
-	unsigned short	digits;
+	unsigned long	digits;
 };
 struct	TdiCatStruct_table	{
 	char		*name;	/*text for decompile*/
-	unsigned short	cat;	/*category code*/
+	unsigned long	cat;	/*category code*/
 	unsigned char	length;	/*size in bytes*/
 	unsigned char	digits;	/*size of text conversion*/
         char            *fname; /*exponent name for floating decompile*/
